+++ frontend/src/components/Theme/ColorModeContext.jsx
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+export const ColorModeContext = createContext({
+  toggleColorMode: () => {},
+});
