+++ frontend/src/components/Theme/index.js
@@ -1,1 +1,2 @@
 export * from './Theme';
+export * from './ColorModeContext';
