+++ frontend/src/hooks/queries/useActiveLoans.jsx
@@ -2,7 +2,7 @@ import { useQuery } from '@tanstack/react-query';
 import { useContext } from 'react';
 
 import { Api } from '../../globals/api';
-import { AlertContext } from '../../providers/AlertProvider';
+import { AlertContext } from '../../providers';
 import { getRequest } from '../../utils';
 
 export const useActiveLoans = (page, options) => {
