+++ frontend/src/providers/AlertContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const AlertContext = createContext();
