+++ frontend/src/providers/CartContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const CartContext = createContext();
