+++ frontend/src/providers/CartProvider.jsx
@@ -1,9 +1,9 @@
 import PropTypes from 'prop-types';
-import { createContext, useState } from 'react';
+import { useState } from 'react';
 
 import { LOCAL_STORAGE_CART_KEY } from '../globals/constants';
 
-export const CartContext = createContext();
+import { CartContext } from './CartContext';
 
 const getCartStateFromLocalStorage = () => {
   const cartItems = JSON.parse(localStorage.getItem(LOCAL_STORAGE_CART_KEY));
