+++ frontend/src/providers/KitCartContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const KitCartContext = createContext();
