+++ frontend/src/providers/KitCartProvider.jsx
@@ -1,9 +1,9 @@
 import PropTypes from 'prop-types';
-import { createContext, useState } from 'react';
+import { useState } from 'react';
 
 import { LOCAL_STORAGE_KIT_CART_KEY } from '../globals/constants';
 
-export const KitCartContext = createContext();
+import { KitCartContext } from './KitCartContext';
 
 const getKitCartItemsFromLocalStorage = () => {
   const cartItems = localStorage.getItem(LOCAL_STORAGE_KIT_CART_KEY);
