+++ frontend/src/providers/NotificationContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const NotificationContext = createContext();
