+++ frontend/src/providers/index.js
@@ -4,3 +4,8 @@ export * from './KitCartProvider';
 export * from './DeckProvider';
 export * from './AuthProvider';
 export * from './NotificationProvider';
+export * from './AlertContext';
+export * from './AuthContext';
+export * from './CartContext';
+export * from './KitCartContext';
+export * from './NotificationContext';
